Added preliminary support for rootfs partition re-sizing
When burning the image to the SD-Card we do loose a lot of space because
only a minimum of space is allocated for the rootfs image. This is caused
by the fact that the image does not know anything about the SD-card size
it will be burned on. The resize-rootfs scripts detects the rootfs partition
and tries to resize the partition.

This may fail when using systemd a bootmanager

Signed-off-by: Christian Ege <k4230r6@gmail.com>