Added MACHINE_FEATURE LVDS7
This enables Qt5.x.x touch support for the KIT LCD 7" Touch device.
To activate this feature add the following to your conf/local.conf

    MACHINE_FEATURES_append = " \
     lvds7 \
    "
This can not be activate as general option because not everybody owns a
7" touch screen panel from UDOO. A more sophisticated could be detecting
the screen based on the current device tree and setting those values
accordingly. This fixes #9

Signed-off-by: Christian Ege <k4230r6@gmail.com>