Fixed loading of M4 firmware on the UDOO Neo boards
In previous releases the M4 firmware was loaded in the u-boot
bootloader. This was changed by the UDOO team for the Ubuntu
image. This moves the firmware loading to userspace like in Ubuntu.

This fixes as-well loading of the uEnv.txt and handling of the
M4 device tree part in u-boot.

Signed-off-by: Christian Ege <k4230r6@gmail.com>