Switched to u-boot-fslc and unified build for all three platforms
* Use u-boot-fslc instead of the UDOO u-boot
* Switched to the latest kernel version
* Use the wic tool for image creation
* Skipped the extra boot partition

The UDOO u-boot changes rapidly which breaks builds. This patch
uses the default u-boot-fslc recipe from meta-fsl-arm plus a small
patch which enables uEnv.txt.

Signed-off-by: Christian Ege <k4230r6@gmail.com>

* Added a bbappend file for the u-boot-fslc recipe

This was missed in the first version

Signed-off-by: Christian Ege <k4230r6@gmail.com>

* Removed old UDOO u-boot entry

This is not needed anymore

* Select the root partition from the boot commandline

This allows booting from eMMc when a SD-Card with
the same partition label is inserted. And we need
no filesystem label while disk creation

Signed-off-by: Christian Ege <k4230r6@gmail.com>

* Get rid of bootlogd error during boot

    bootlogd: cannot allocate pseudo tty: No such file or directory

Signed-off-by: Christian Ege <k4230r6@gmail.com>

* Added WIC kickstarter file for image creation

The wic tool provides a flexible image creation process
To generate the image run the following

    wic create udoo-sdcard -e udoo-image-full-cmdline -o /var/tmp/wic/udooneo

Signed-off-by: Christian Ege <k4230r6@gmail.com>

* Provide information how to run the wic tool

Signed-off-by: Christian Ege <k4230r6@gmail.com>

* Rebased the Patch set from Francesco Montefoschi

This is based on the u-boot provided by Francesco Montefoschi

https://github.com/fmntf/u-boot

* Switched to the new u-boot aproach

Signed-off-by: Christian Ege <k4230r6@gmail.com>

* Fixed loading of M4 firmware on the UDOO Neo boards

In previous releases the M4 firmware was loaded in the u-boot
bootloader. This was changed by the UDOO team for the Ubuntu
image. This moves the firmware loading to userspace like in Ubuntu.

This fixes as-well loading of the uEnv.txt and handling of the
M4 device tree part in u-boot.

Signed-off-by: Christian Ege <k4230r6@gmail.com>

* Unified the UDOO and Seco A62 boards

Now both use the same u-boot and kernel repositories

Signed-off-by: Christian Ege <k4230r6@gmail.com>

* Added missing environment files

Signed-off-by: Christian Ege <k4230r6@gmail.com>

* Added missing files

Signed-off-by: Christian Ege <k4230r6@gmail.com>

* Corrects the SPI and CAN buses patches.

Signed-off-by: Cyril Franceschini <cyril.franceschini@eeproperty.ch>
Signed-off-by: Christian Ege <k4230r6@gmail.com>

* Removed obsolete stuff

Signed-off-by: Christian Ege <k4230r6@gmail.com>